requirements.txt
awscli==1.16.258
boto3==1.9.184
botocore==1.12.248
bugsnag==3.6.0
beautifulsoup4==4.8.1
docutils==0.14
gql==0.1.0
graphql-core==2.2.1
numpy==1.16.4
pandas==0.24.2
promise==2.2.1
psycopg2-binary==2.8.4
pyarrow==0.14.0
PyInstaller==3.5
python-dateutil==2.8.0
python-dotenv==0.10.3
pytz==2019.1
PyYAML==5.1.2
requests==2.22.0
s3fs==0.2.2
six==1.12.0
SQLAlchemy==1.3.12
urllib3==1.25.3